+++ R/canvas_flow.R
@@ -3,7 +3,7 @@
 #' @description This function draws flow fields on a canvas.
 #'
 #' @usage canvas_flow(colors, background = "#fafafa", lines = 500, lwd = 0.05,
-#'             iterations = 100, stepmax = 1, angles = NULL)
+#'             iterations = 100, stepmax = 0.01, angles = NULL)
 #'
 #' @param colors         a string or character vector specifying the color(s) used for the artwork.
 #' @param background     a character specifying the color used for the background.