+++ R/colorPalette.R
@@ -4,8 +4,8 @@
 #'
 #' @usage colorPalette(name, n = NULL)
 #'
-#' @param name   name of the color palette. Can be \code{random} for random colors, but can also be the name of a pre-implemented palette. See the \code{details} section for a list of pre-implemented palettes.
-#' @param n      the number of colors to select from the palette. Required if \code{name = 'random'}. Otherwise, if \code{NULL}, automatically selects all colors from the chosen palette.
+#' @param name   name of the color palette. Can be \code{random} for random colors or \code{complement} for complementing colors, but can also be the name of a pre-implemented palette. See the \code{details} section for a list of pre-implemented palettes.
+#' @param n      the number of colors to select from the palette. Required if \code{name = 'random'} or \code{name = 'complement'}. Otherwise, if \code{NULL}, automatically selects all colors from the chosen palette.
 #'
 #' @details The following color palettes are implemented:
 #'
@@ -17,7 +17,7 @@
 #' @author Koen Derks, \email{koen-derks@hotmail.com}
 #'
 #' @examples
-#' colorPalette("random", 5)
+#' colorPalette("complement", 5)
 #' @keywords canvas palette
 #'
 #' @export
@@ -33,6 +33,23 @@ colorPalette <- function(name, n = NULL) {
     palette <- character(n)
     for (i in 1:length(palette)) {
       palette[i] <- .hsl_to_rgb(h = stats::runif(1, 0, 360), stats::runif(1), stats::runif(1))
+    }
+  } else if (name == "complement") {
+    palette <- character(n)
+	tmp <- stats::runif(1, 0, 360)
+	palette[1] <- .hsl_to_rgb(h = tmp, stats::runif(1), stats::runif(1))
+    for (i in 2:length(palette)) {
+      if (i%%2 == 0) {
+		  if (tmp >= 180) {
+			  color <- tmp - 180
+		  } else {
+			  color <- tmp + 180
+		  }
+		  palette[i] <- .hsl_to_rgb(h = color, stats::runif(1, .4, 1), stats::runif(1, .4, 1))
+	  } else {
+		  tmp <- stats::runif(1, 0, 360)
+		  palette[i] <- .hsl_to_rgb(h = tmp, stats::runif(1, .4, 1), stats::runif(1, .4, 1))
+	  }
     }
   } else {
     palette <- switch(name,